+++ bin/update-readmes.js
@@ -13,7 +13,7 @@ const packages = [
 	'block-serialization-default-parser',
 	'blocks',
 	'compose',
-	//'data',
+	'data',
 	'date',
 	'deprecated',
 	'dom',