+++ packages/block-library/src/latest-posts/index.php
@@ -89,10 +89,14 @@ function register_block_core_latest_posts() {
 		'core/latest-posts',
 		array(
 			'attributes'      => array(
-				'categories'      => array(
+				'align'           => array(
 					'type' => 'string',
+					'enum' => array( 'left', 'center', 'right', 'wide', 'full' ),
 				),
 				'className'       => array(
+					'type' => 'string',
+				),
+				'categories'      => array(
 					'type' => 'string',
 				),
 				'postsToShow'     => array(
@@ -110,9 +114,6 @@ function register_block_core_latest_posts() {
 				'columns'         => array(
 					'type'    => 'number',
 					'default' => 3,
-				),
-				'align'           => array(
-					'type' => 'string',
 				),
 				'order'           => array(
 					'type'    => 'string',