+++ packages/block-library/src/search/index.js
@@ -21,5 +21,9 @@ export const settings = {
 
 	keywords: [ __( 'find' ) ],
 
+	supports: {
+		align: true,
+	},
+
 	edit,
 };